Jeg har opdateret GUI_fields til at caste klasserne på GUI
